# src/Otus.Highload.Api/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "OTUS Highload Architect",
    "version" : "1.2.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/user/register" : {
      "post" : {
        "description" : "Регистрация нового пользователя",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/user_register_body"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Успешная регистрация",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "400" : {
            "description" : "Невалидные данные"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        }
      }
    },
    "/user/get/{id}" : {
      "get" : {
        "description" : "Получение анкеты пользователя",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Идентификатор пользователя",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UserId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Успешное получение анкеты пользователя",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Невалидные данные"
          },
          "404" : {
            "description" : "Анкета не найдена"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        }
      }
    },
    "/user/search" : {
      "get" : {
        "description" : "Поиск анкет",
        "parameters" : [ {
          "name" : "first_name",
          "in" : "query",
          "description" : "Условие поиска по имени",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "description" : "Часть имени для поиска",
            "example" : "Конст"
          }
        }, {
          "name" : "last_name",
          "in" : "query",
          "description" : "Условие поиска по фамилии",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "description" : "Часть фамилии для поиска",
            "example" : "Оси"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Успешные поиск пользователя",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Невалидные данные"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "BirthDate" : {
        "type" : "string",
        "description" : "Дата рождения",
        "format" : "date",
        "example" : "2017-02-01"
      },
      "UserId" : {
        "type" : "string",
        "description" : "Идентификатор пользователя"
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UserId"
          },
          "first_name" : {
            "type" : "string",
            "description" : "Имя",
            "example" : "Имя"
          },
          "second_name" : {
            "type" : "string",
            "description" : "Фамилия",
            "example" : "Фамилия"
          },
          "birthdate" : {
            "$ref" : "#/components/schemas/BirthDate"
          },
          "biography" : {
            "type" : "string",
            "description" : "Интересы",
            "example" : "Хобби, интересы и т.п."
          },
          "city" : {
            "type" : "string",
            "description" : "Город",
            "example" : "Москва"
          }
        },
        "example" : {
          "birthdate" : "2017-02-01T00:00:00.000+00:00",
          "city" : "Москва",
          "second_name" : "Фамилия",
          "id" : "id",
          "biography" : "Хобби, интересы и т.п.",
          "first_name" : "Имя"
        }
      },
...
      "inline_response_500" : {
        "required" : [ "message" ],
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Описание ошибки"
          },
          "request_id" : {
            "type" : "string",
            "description" : "Идентификатор запроса. Предназначен для более быстрого поиска проблем."
          },
          "code" : {
            "type" : "integer",
            "description" : "Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем."
          }
        }
      },
      "user_register_body" : {
        "type" : "object",
        "properties" : {
          "first_name" : {
            "type" : "string",
            "example" : "Имя"
          },
          "second_name" : {
            "type" : "string",
            "example" : "Фамилия"
          },
          "birthdate" : {
            "$ref" : "#/components/schemas/BirthDate"
          },
          "biography" : {
            "type" : "string",
            "example" : "Хобби, интересы и т.п."
          },
          "city" : {
            "type" : "string",
            "example" : "Москва"
          },
          "password" : {
            "type" : "string",
            "example" : "Секретная строка"
          }
        }
      },
      "inline_response_200_1" : {
        "type" : "object",
        "properties" : {
          "user_id" : {
            "type" : "string",
            "example" : "e4d2e6b0-cde2-42c5-aac3-0b8316f21e58"
          }
        },
        "example" : {
          "user_id" : "e4d2e6b0-cde2-42c5-aac3-0b8316f21e58"
        }
      },
...
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Авторизация по токену, который был получен в методе /login",
        "scheme" : "bearer"
      }
    }
  }
}